refactor: changed files name spelling
