+++ src/shared/types/index.ts
@@ -1,1 +1,1 @@
-export * from './user.interface';
+export * from './userInterface';
